requirements.txt
pandas
psycopg2-binary
fastapi
uvicorn[standard]
asyncpg
